Alright, Naia's pretty much working now
If you approach her, you get the option to talk to her. If you want the option to go away, just press enter. If you want the option to come back, press enter again. If you start fighting with her, you can't get the option to talk to her anymore.